% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{edl_extract_urls}
\alias{edl_extract_urls}
\title{data access URLs}
\usage{
edl_extract_urls(resp)
}
\arguments{
\item{resp}{a response object from edl_search}
}
\value{
a character vector of URLs
}
\description{
data access URLs
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

resp <- edl_search(short_name = "MUR-JPL-L4-GLOB-v4.1",
                   temporal = c("2020-01-01", "2021-12-31"))

urls <- edl_extract_urls(resp)
\dontshow{\}) # examplesIf}
}
