% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edl.R
\name{edl_header}
\alias{edl_header}
\title{edl_header}
\usage{
edl_header(token = edl_set_token())
}
\arguments{
\item{token}{the authentication token, e.g. from \code{edl_set_token}}
}
\description{
Return the http authentication header for curl-based requests
}
