% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login-helpers.R
\name{edl_renviron}
\alias{edl_renviron}
\title{Renviron Helper utility}
\usage{
edl_renviron(user = "", password = "", verbose = TRUE)
}
\arguments{
\item{user}{your EDL username (created at
\url{https://urs.earthdata.nasa.gov/home})}

\item{password}{your EDL password (created at
\url{https://urs.earthdata.nasa.gov/home})}

\item{verbose}{show messages?}
}
\value{
adds environmental variables to .Renviron and returns
NULL (invisibly.)
}
\description{
Asks user for their username and password and store the result in Rprofile.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
edl_renviron()
\dontshow{\}) # examplesIf}
}
