% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edl_set_token.R
\name{edl_stac_urls}
\alias{edl_stac_urls}
\title{Helper function for extracting URLs from STAC}
\usage{
edl_stac_urls(items, assets = "data")
}
\arguments{
\item{items}{an items list from rstac}

\item{assets}{name(s) of assets to extract}
}
\value{
a vector of hrefs for all discovered assets.
}
\description{
Helper function for extracting URLs from STAC
}
