% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edl_s3_token.R
\name{edl_unset_s3}
\alias{edl_unset_s3}
\title{Unset AWS S3 Environment Variables}
\usage{
edl_unset_s3()
}
\description{
The function uses \code{Sys.unsetenv()} to remove the specified environment variables.
}
\details{
This function unsets the AWS S3-related environment variables:
\code{AWS_ACCESS_KEY_ID}, \code{AWS_SECRET_ACCESS_KEY}, and \code{AWS_SESSION_TOKEN}.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
edl_unset_s3()
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link{Sys.unsetenv}}
}
