% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edl.R
\name{edl_unset_token}
\alias{edl_unset_token}
\title{unset token}
\usage{
edl_unset_token()
}
\description{
External sources that don't need the token may error if token is set.
Call \code{edl_unset_token} before accessing non-EarthData URLs.
}
\examples{
edl_unset_token()
}
