% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list-nasa-stacs.R
\name{list_nasa_stacs}
\alias{list_nasa_stacs}
\title{List NASA STAC Catalogs}
\usage{
list_nasa_stacs(cloud_only = TRUE)
}
\arguments{
\item{cloud_only}{Logical; if TRUE (default), returns only cloud-hosted STAC catalogs}
}
\value{
A data frame of STAC catalog urls for all STAC providers in the CMR
}
\description{
Retrieves available STAC catalogs from NASA's Common Metadata Repository (CMR)
}
\examples{
\dontrun{
list_nasa_stacs()
list_nasa_stacs(cloud_only = FALSE)
}

}
\seealso{
\code{\link[=get_nasa_stac_url]{get_nasa_stac_url()}}
}
