% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edl.R
\name{lpdacc_example_url}
\alias{lpdacc_example_url}
\title{URL for an example of an LP DAAC COG file}
\usage{
lpdacc_example_url()
}
\value{
The URL to a Cloud-Optimized Geotiff file from the LP DAAC.
}
\description{
URL for an example of an LP DAAC COG file
}
\examples{
lpdacc_example_url()

}
