% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{edl_extract_urls}
\alias{edl_extract_urls}
\title{data access URLs}
\usage{
edl_extract_urls(resp)
}
\arguments{
\item{resp}{a response object from edl_search}
}
\value{
a character vector of URLs
}
\description{
data access URLs
}
