% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edl.R
\name{edl_s3_token}
\alias{edl_s3_token}
\title{Receive and set temporary AWS Tokens for S3 access}
\usage{
edl_s3_token(
  daac = "https://data.lpdaac.earthdatacloud.nasa.gov",
  username = default("user"),
  password = default("password")
)
}
\arguments{
\item{daac}{the base URL for the DAAC}

\item{username}{EarthDataLogin user}

\item{password}{EarthDataLogin Password}
}
\value{
list of access key, secret key, session token and expiration,
invisibly.  Also sets the corresponding AWS environmental variables.
}
\description{
Receive and set temporary AWS Tokens for S3 access
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
edl_s3_token()
\dontshow{\}) # examplesIf}
}
