% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edl.R
\name{edl_set_token}
\alias{edl_set_token}
\title{Get or set an earthdata login token}
\usage{
edl_set_token(
  username = Sys.getenv("EARTHDATA_USER"),
  password = Sys.getenv("EARTHDATA_PASSWORD"),
  token_number = 1
)
}
\arguments{
\item{username}{EarthData Login User}

\item{password}{EarthData Login Password}

\item{token_number}{Which token (1 or 2)}
}
\value{
the Authorization header for curl request, invisibly.
Will also set the GDAL_HTTP_HEADERS environmental variable for
compatibility with GDAL.
}
\description{
This function will ping the EarthData API for any available tokens.
If a token is not found, it will request one.  You may only have
two active tokens at any given time.  Use edl_revoke_token to remove
unwanted tokens.
}
\details{
NOTE: GDAL >= 3.6 is required to utilize GDAL_HTTP_HEADERS.
}
